Usar tipo real en programa máximo reales

+++ diapos/programas/mayor-reales.py
@@ -1,6 +1,6 @@
 mayor = -float('inf')
 while True:
-    n = int(raw_input())
+    n = float(raw_input())
     if n == 0:
         break
     mayor = max(mayor, n)
